// src/app/models/validation/insurnace.comp.info.required.ts
export interface InsurnaceCompInfoRequired {
    compNoFault: boolean;
    relatedInjury: boolean;
    accidentDate: boolean
    workerStatus: boolean
    address: boolean;
    fax: boolean
    insuranceName: boolean
    claimNumber: boolean;
    adjusterName: boolean
    adjusterPhone: boolean
    attorneyName: boolean
    attorneyPhone: boolean
    caseStatus: boolean
}